test(tau-coding-agent): split tooling_skills remote install tests (#1385)
